+++ src/torchfem/base.py
@@ -9,7 +9,7 @@ import torch
 from torch import Tensor
 
 from .elements import Element
-from .materials import Material
+from .materials import HeatMaterial, Material, MechanicsMaterial
 from .report import solve_report
 from .sparse import (
     differentiable_modal_eigsolve,
@@ -981,7 +981,7 @@ class Mechanics(FEM, ABC):
         flux_new = torch.zeros_like(flux_prev)
         state_new = torch.zeros_like(state_prev)
 
-        assert self.material is not None
+        assert isinstance(self.material, MechanicsMaterial)
 
         # Compute gradient operators
         _, B, detJ = self.eval_shape_functions(self.etype.ipoints)
@@ -1159,7 +1159,7 @@ class Heat(FEM, ABC):
         f = torch.zeros(self.n_elem, n_dof, device=du.device)
         k = torch.zeros(self.n_elem, n_dof, n_dof, device=du.device) if need_k else None
 
-        assert self.material is not None
+        assert isinstance(self.material, HeatMaterial)
 
         grad_new = []
         flux_new = []
@@ -1180,7 +1180,6 @@ class Heat(FEM, ABC):
                 grad_prev[i],
                 flux_prev[i],
                 state_prev[i],
-                de0,
                 self.char_lengths,
                 iter,
             )